requirements.txt
scipy==1.1.0
torch
SoundFile==0.10.2
sacred==0.7.4
librosa==0.6.2
numpy
tqdm
mir_eval
mido==1.2.9
Pillow==5.3.0
tensorboardX==1.2
tensorflow
